requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==0.8.11
requests==2.31.0
gunicorn==21.2.0
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
